Use existing IDE archives instead of downloading them if they already exist

+++ install.sh
@@ -6,15 +6,23 @@ su - vagrant
 export DEBIAN_FRONTEND=noninteractive
 
 clion_version=CLion-2016.1-1
+ide_archive_location=/home/vagrant/ide-archive
 
 touch /home/vagrant/.bash_profile && \
-(if grep -v -q startx /home/vagrant/.bash_profile; then echo "[[ -z $DISPLAY && $XDG_VTNR -eq 1 ]] && exec startx" >> /home/vagrant/.bash_profile; fi) && \
+( \
+ if grep -v -q startx /home/vagrant/.bash_profile; then \
+  echo "[[ -z $DISPLAY && $XDG_VTNR -eq 1 ]] && exec startx" >> /home/vagrant/.bash_profile; \
+ fi \
+) && \
 \
 # Adding package sources for NodeJS and Java 8. The NodeJS script executes apt-get update as well
 curl -sL https://deb.nodesource.com/setup_4.x | bash - && \
 echo 'oracle-java8-installer shared/accepted-oracle-license-v1-1 boolean true' | debconf-set-selections && \
-(if [ ! -f /etc/apt/sources.list.d/webupd8team-java.list ]; then \
-printf 'deb http://ppa.launchpad.net/webupd8team/java/ubuntu trusty main\n\ndeb-src http://ppa.launchpad.net/webupd8team/java/ubuntu trusty main' >> /etc/apt/sources.list.d/webupd8team-java.list; fi) && \
+( \
+ if [ ! -f /etc/apt/sources.list.d/webupd8team-java.list ]; then \
+  printf 'deb http://ppa.launchpad.net/webupd8team/java/ubuntu trusty main\n\ndeb-src http://ppa.launchpad.net/webupd8team/java/ubuntu trusty main' >> /etc/apt/sources.list.d/webupd8team-java.list; \
+ fi \
+) && \
 # Installing NodeJS
 apt-get -y install nodejs && \
 \
@@ -68,7 +76,12 @@ apt-get -y autoremove && \
 # Downloading and extracting CLion
 (if [ ! -d ${clion_version} ]; \
  then ( \
-	wget http://download.jetbrains.com/cpp/${clion_version}.tar.gz && \
+  ( \
+   mv ${ide_archive_location}/* . 2>/dev/null &&
+   if [ ! -f ${clion_version}.tar.gz ]; \
+	  then wget https://download.jetbrains.com/cpp/${clion_version}.tar.gz
+	 fi \
+	) && \
 	tar -xvzf ${clion_version}.tar.gz && \
 	chown vagrant:vagrant ${clion_version} && \
 	rm ${clion_version}.tar.gz; \
